refactor(web): implement joined user fetching with custom hook in `ProjectHeader`
