+++ apps/web/src/features/project/api/index.ts
@@ -1,3 +1,2 @@
 export * from './invite';
-export * from './joinedUser';
 export * from './update';
