+++ apps/web/src/widgets/project/api/index.ts
@@ -1,1 +1,2 @@
 export * from './projectBoard';
+export * from './joinedUser';
